Allow all types for IIdentifiableObject Fix ReadOnly Attribute
