+++ Demo/ReadOnlyDemo.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dced5097e38b476e82ca519db2defd7d, type: 3}
+  m_Name: ReadOnlyDemo
+  m_EditorClassIdentifier: 
+  uuid: 3c82ab46-e6d5-4535-abd1-1543f17d38e8
+  opt: hjzujhj
+  optChange: 1
