+++ Demo/ReadOnlyDemo.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bcac9c1bb5d721247bc5d5d431971a07
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
