+++ Scripts/Demo/UnityEditorEx.Demo.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5dbc31d7d8b1f254ca77bece0d12e726
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
