+++ Scripts/Runtime/Types/IIdentifiableObject.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace UnityEditorEx.Runtime.Projects.unity_editor_ex.Scripts.Runtime.Types
+{
+    public interface IIdentifiableObject<out T>
+    {
+        T Identifier { get; }
+    }
+}
