+++ Scripts/Runtime/Types/IIdentifiableObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc6057b5e79ec9648974ef1437c33d16
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
