--- Scripts/Runtime/Types/IIdentifiedObject.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace UnityEditorEx.Runtime.Projects.unity_editor_ex.Scripts.Runtime.Types
-{
-    public interface IIdentifiedObject<out T> where T : Enum
-    {
-        T Identifier { get; }
-    }
-}
