+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "org.pcsoft.editor-ex",
-  "version": "1.2.2",
+  "version": "1.3.0",
   "displayName": "Unity Editor Extension",
   "description": "Extension for Editor system of Unity.",
   "unity": "2021.3",
